Making delay input float for WorkerB and WorkerC.

+++ worker_c/main.py
@@ -12,7 +12,7 @@ PROJECT_ID = os.environ.get('PROJECT_ID')
 PUBLISHER_SUCCESS_TOPIC_ID = os.environ.get('PUBLISHER_SUCCESS_TOPIC_ID')
 PUBLISHER_ERROR_TOPIC_ID = os.environ.get('PUBLISHER_ERROR_TOPIC_ID')
 NODE_ID = int(os.environ.get('NODE_ID', 3))
-DELAY = int(os.environ.get('DELAY', 15))
+DELAY = float(os.environ.get('DELAY', 15))
 
 publisher = pubsub_v1.PublisherClient()
